CHROMIUM: pm-check: Don't check / bitfix invalid pages
It turns out that there are some chunks of memory that are "removed"
from the system at boot.  These are:
  0x43000000 (0x00800000 bytes) - MFC
  0x51000000 (0x00800000 bytes) - MFC
  0x41f00000 (0x00100000 bytes) - pstore

It turns out that if we're not using SPARSEMEM_EXTREME that it's not so
valid to call kmap_atomic() on these physical addresses.  Doing so will
return a virtual address that causes a crash when accessed.

This memory isn't checksumed by pm-check anyway.  If you look at the
pm-check printouts, you see:
   s3c_pm_check: Found system RAM at 40000000..41efffff
   s3c_pm_check: Found system RAM at 42000000..42ffffff
   s3c_pm_check: Found system RAM at 43800000..50ffffff
   s3c_pm_check: Found system RAM at 51800000..bfffffff
   s3c_pm_check: Found system RAM at 40000000..41efffff

What matters, though, is that bitfix wasn't excluding these chunks of
memory during its XOR calculations.  Bitfix relied on pm_check to tell
it which chunks of memory to skip and pm_check wasn't including these in
the list of things to skip.

It was OK in the past for bitfix to use this memory since it was in the
XOR both before and after suspend and it wasn't changing during
suspend.  ...but not including it shouldn't be terrible since we weren't
checksumming it anyway.

Once we fix this, we can turn off SPARSEMEM_EXTREME and in turn turn on
compaction (which was broken with SPARSEMEM_EXTREME).

From testing, pfn_valid() is a proper way to exclude these chunks on our
3.8 exynos kernel.  Presumably we could also iterate over all the
resources looking for memory resources, as was done elsewhere in
pm-check.

BUG=chromium:586052
TEST=suspend_stress_test with CL:Idafb48f4fd56cfb65119271a60604cc420223a44

Change-Id: I9488c497383263f6de139f9df0168a0d69109687
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/327481
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>