CHROMIUM: input: elan_i2c - Append PID to fw names
Previously, the elan_i2c touchpad driver in the 3.8 and 3.10 kernel had
two ways of triggering it.
  1. You "cat" a 1 into the update_fw sysfs entry.  When you do this it
     simply loaded the default /lib/firmware/elan_i2c.bin and uses that
     as the FW to update with.
  2. You "cat" a filename into updat_fw.  When you do this, it used
     /lib/firmware/${WHAT_YOU_CATTED} instead.

The more modern versions of this driver (kernel >= 3.14) do not offer
option #2 anymore.  Instead, regardless of what you cat in they will
first try to load /lib/firmware/elan_i2c_${PRODUCT_ID}.bin, and then if
that is not found, will fail back to /lib/firmware/elan_i2c.bin in an
effort to maintain backwards capability.

The logic behind adding the product_id onto the end of the filename is
to allow a single device/image to multi-source the touchpads while
maintaining different FWs for each one.

eg: if we have two Elan touchpads in one laptop with product_id's 1.0
and 2.0, then we just have to store a elan_i2c_1.0.bin and an
elan_i2c_2.0.bin in /lib/firmware and the updater will always use the
correct one.

We never had this functionality in 3.8 or 3.10, however.  This CL adds
the product_id appending code to this version of the driver (failing
back to elan_i2c.bin for backwards compatibility)  to allow devices
using this older version to support multiple Elan touchpads with
separate firmwares.

BUG=chrome-os-partner:50932
TEST=emerged onto a Peppy, and tested manually on it
CQ-DEPEND=CL:*253672

Change-Id: Iddac01749eff24c55ad35c9e4a648eefb455aad4
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/338905
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>