CHROMIUM: iwl7000: mac80211: linearize RX frame before extracting 802.3s
The upstream version of cfg80211 (i.e. >= v4.7) now supports paging in
the skbuff when extracting 802.3 packets from an AMSDU.

The mac80211 code delivered with iwl7000 followed upstream, where the
linearization is not needed anymore.  But chromeos's vesion of
cfg80211 doesn't have this change yet, so mac80211 still needs to
linearize before passing the buffer to cfg80211.

BUG=chrome-os-partner:57003
TEST=connect to an AP (e.g. OnHub) that sends several AMPDUs embedded
in an AMSDU.

Change-Id: I652c2a56bd407d71ae6d35aab0e2a28bfb7c09f6
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/381159
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>