CHROMIUM: iwl7000: mvm: fix reserved txq freeing
If a TXQ's marking as a reserved queue is removed,
when removing the STA the driver might try to access
out of bounds memory. Make sure the reserved queue
is freed only if it is still reserved.

Signed-off-by: Liad Kaufman <liad.kaufman@intel.com>
iwl7000-tree: 7522eb83b7f8a9f9792b342f7bfe19c4f776db89