CHROMIUM: iwl7000: mvm: let the FW choose its antenna for the beacon
I added the support for letting the firmware choose the
antenna for the beacon, I thought the firmware would
override the driver's settings in the Beacon's Tx command
but it doesn't.
Don't set any antenna in the Tx command of the Beacon
command to let the firmware which antenna it wants.

Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 6b28057c02fdb0e7f6d695f52c57d26f1811bfc2