CHROMIUM: iwl7000: mvm: workaround for a FW bug and pick the beacon transmission's antenna
The firmware is currently buggy and can't handle the case
in which we let the firmware choose the antenna to use for
the beacon transmission. As a workaround, always choose the
antenna in the driver until the bug is fixed in the
firmware.

Signed-off-by: Emmanuel Grumbach <emmanuel.grumbach@intel.com>
iwl7000-tree: 6ef53bb387f6d2350ca3c532962f77cba4a5cee2