CHROMIUM: iwl7000: mac80211: set wifi_ackedbits for transmitted SKBs
There may be situations in which the in-kernel originator of an
SKB cares about its wifi transmission status. To have that, set
the wifi_acked[_valid] bits before freeing/orphaning the SKB if
the destructor is set. The originator can then use it in there.

Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 049a9d355c1e80a3536bba5285be8648d0a80643